Adds tox configuration
Local testing is lacking the same environment support as in Travis.
This translates the current .travis.yml into a tox configuration.

The user can just run `tox` and test building, packaging and several
interpreters with one stroke.

The `setup.py sdist` step has been omitted as it is part of the tox default
behaviour.